--- templateSpring/templateSpring/src/main/java/com/Template/templateSpring/TemplateSpringApplication.java
@@ -1,12 +1,0 @@
-package com.Template.templateSpring;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TemplateSpringApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(TemplateSpringApplication.class, args);
-	}
-
-}
