--- templateSpring/templateSpring/src/main/java/com/Template/templateSpring/dto/UserSignUpDto.java
@@ -1,46 +1,0 @@
-package com.Template.templateSpring.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
-
-public class UserSignUpDto {
-
-    @NotEmpty
-    @NotBlank
-    private String email;
-    @NotEmpty
-    @NotBlank
-    private String password;
-
-    public void setConfirmationPassword(String confirmationPassword) {
-        this.confirmationPassword = confirmationPassword;
-    }
-
-    private String confirmationPassword;
-
-    public UserSignUpDto() {
-
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getConfirmationPassword() {
-        return confirmationPassword;
-    }
-
-
-}
