--- templateSpring/templateSpring/src/main/java/com/Template/templateSpring/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.Template.templateSpring.repository;
-
-import com.Template.templateSpring.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    User findByEmail(String email);
-    Boolean existsByEmail(String email);
-}
