--- templateSpring/templateSpring/src/main/java/com/Template/templateSpring/service/UserService.java
@@ -1,8 +1,0 @@
-package com.Template.templateSpring.service;
-
-import com.Template.templateSpring.dto.UserSignUpDto;
-
-public interface UserService {
-    void saveUser(UserSignUpDto userSignUpDto);
-
-}
