+++ templateSpring/templateSpring/src/main/java/tech/aluve/calendar/repository/UserRepository.java
@@ -1,0 +1,9 @@
+package tech.aluve.calendar.repository;
+
+import tech.aluve.calendar.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findByEmail(String email);
+    Boolean existsByEmail(String email);
+}
