+++ templateSpring/templateSpring/src/main/java/tech/aluve/calendar/service/UserService.java
@@ -1,0 +1,8 @@
+package tech.aluve.calendar.service;
+
+import tech.aluve.calendar.dto.UserSignUpDto;
+
+public interface UserService {
+    void saveUser(UserSignUpDto userSignUpDto);
+
+}
