+++ templateSpring/templateSpring/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 #### PostgresSql Properties ####
 spring.datasource.url=jdbc:postgresql://localhost:5432/aluvedb
-spring.datasource.username=${postuser}
-spring.datasource.password=${postpass}
+spring.datasource.username=mzwilipost
+spring.datasource.password=mzwili@post
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
@@ -15,3 +15,11 @@ spring.mail.username=${email}
 spring.mail.password=${emailpassword}
 spring.mail.properties.mail.smtp.auth=true
 spring.mail.properties.mail.smtp.starttls.enable=true
+
+##### logging properties ####
+#logging.level.com.dailycodebuffer = TRACE
+#
+#
+#logging.pattern.console=%d [%level] %c{1.} [%t] %m%n
+#logging.file = appLog.log
+#logging.pattern.file=%d [%level] %c{1.} [%t] %m%n
