--- templateSpring/templateSpring/src/test/java/com/Template/templateSpring/TemplateSpringApplicationTests.java
@@ -1,13 +1,0 @@
-package com.Template.templateSpring;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TemplateSpringApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
