+++ templateSpring/templateSpring/src/test/java/tech/aluve/calendar/CalendarAppTests.java
@@ -1,0 +1,13 @@
+package tech.aluve.calendar;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CalendarAppTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
